Add message to users after 24 hours
Due to lack of user engagement with the product, it was decided to simply send the users a message after 24 hours if they failed to launch a game properly. The message asks the user if they need any help, so that we can understand what the difficulty is. To do this I added EXQ.

Some bugs were also fixed on the notifications for sending long messages.