Add index to games for public games
The boolean public was added to the games table, so now users can host or join games without being friends.

It was decided that public games should not be included in the me function, and would have to be fetched by the controller instead. The boolean public is used to differentiate between the to types.