Remove public boolean, replace with join_code
All games are now planned to be shown on the games tab. Some of those games may be locked by a code if they want only friends/family to join. This change adds the join_code to the games table and removes the public boolean, as no longer planned for use.