Add ability to remove player who already joined game
Since public games will allow users to be booted or to leave the game, the functionality needed to be added to exit the game. The delete function on the invitations controller allows a user to be removed from game. If host is the current user, any user in the game can be removed, if a standard user, then only that user can remove themselves.