As notify_on_game_creation boolean to users
Adding a new boolean, intended for use by new users. To avoid users joining and no game be available, we will notify them when another user makes a new game. This boolean represents whether they want to receive this notification or not, but for new users, we will assume it is true.