Send new game creation notification on game creation
In order to try and bring users back who were looking for an active game but unable to find one, a notification will be sent. Hopefully we will get enough users returning for each notification that a game and community will eventually start to grow. If that does happen, then we can turn off the notification.