Game index only shows game which are less than 24 hours old
The app had a problem where people keep joining games which had the most players, but the host had long since left, giving the users a frustrating experience, as they wondered why the game wasn't starrting. This is a step in the right direction to solving it, as now only newer games will be shown, where the host is less likely to be inactive.

But it is only a small step in the right direction...