Add type to bot messages
We plan to replace bot messages at some point in the future with unique widgets in the app, this gives us the flexibility to do that.

Also a pre-step to our first bot message widget, asking the users to join our discord server on either 24 hours with no game start, or after a successful game.