Add uuid and destinaiton to messages
UUID is need so that the front end can automatically add a message to the display on the app without worrying about duplicating the message on the screen.

Destination is added to the game_messages table so that we can seperate messages on the game into four or more types in the future, standard, werewolf, mason and dead. This is so that we can seperate the game chat into seperate tabs. This is just the ground work for that work.