Announce all game messages to the game itself
This follows the change where we now have more than one tab for chatting on the app when playing a game. Now all announcement messages are sent to the game itself, with different destination strings.

The original conversation announcements will be removed following a large % of users updating the app.