Allow whole game to be updated on update
Whenever the update function on the game controller is called, we call edit on the game itself, to ensure the game is always updated.

We also now update the whole game in the database as well, including the name and join code.

This should give a much better user experience, as they no longer have to leave a stale incorrect game if they made a mistake.