Added phase number to game message
This will be used for custom end of phase and game over messages to replace the text based message from the server. We needed the phase number so we could know which data to get from the state for each specific end of phase message. It is not important that regular messages have a phase_number of zero for now.